+++ app/src/main/res/layout/item_hotline.xml
@@ -30,9 +30,8 @@
             android:textAlignment="textStart"
             android:textColor="@color/Indigo"
 
-            android:textSize="24sp"
-            android:textStyle="bold"
-            app:layout_constraintEnd_toEndOf="parent"
+            style="@style/Header4"
+            app:layout_constraintEnd_toStartOf="@+id/phoneNumber"
             app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="@+id/hotlineHeader"
 
@@ -44,8 +43,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textColor="@color/Greyish_Brown"
-            android:textSize="16sp"
-            android:textStyle="bold"
+            style="@style/body1"
 
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.133"
